Fixed spelling mistakes in in-code documentation Moved `migrator` to `RootDirectory` Internal structure reorganisation
