+++ example/lib/main.dart
@@ -21,7 +21,7 @@ void main() async {
   final SharedPreferences prefs = await SharedPreferences.getInstance();
 
   FlutterMapTileCaching.initialise(await RootDirectory.normalCache);
-  await FMTC.instance.migrator.fromV4();
+  await FMTC.instance.rootDirectory.migrator.fromV4();
 
   if (prefs.getBool('reset') ?? false) {
     await FMTC.instance.rootDirectory.manage.resetAsync();
