+++ example/lib/screens/download_region/components/section_separator.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class SectionSeperator extends StatelessWidget {
-  const SectionSeperator({Key? key}) : super(key: key);
+class SectionSeparator extends StatelessWidget {
+  const SectionSeparator({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) => Column(
