+++ lib/fmtc_advanced.dart
@@ -11,6 +11,7 @@ export 'src/bulk_download/progress_management.dart';
 export 'src/bulk_download/tile_loops.dart';
 export 'src/bulk_download/tile_progress.dart';
 export 'src/internal/exts.dart';
+export 'src/internal/filesystem_sanitiser_private.dart';
 export 'src/internal/image_provider.dart';
 export 'src/internal/recovery/decode.dart';
 export 'src/internal/recovery/encode.dart';
@@ -19,7 +20,6 @@ export 'src/root/access.dart';
 export 'src/root/manage.dart';
 export 'src/root/recovery.dart';
 export 'src/root/statistics.dart';
-export 'src/settings/filesystem_sanitiser_private.dart';
 export 'src/store/access.dart';
 export 'src/store/download.dart';
 export 'src/store/manage.dart';
