+++ lib/src/regions/downloadable_region.dart
@@ -29,7 +29,7 @@ class DownloadableRegion {
   /// The original [BaseRegion], used internally for recovery purposes
   final BaseRegion originalRegion;
 
-  /// All the verticies on the outline of a polygon
+  /// All the vertices on the outline of a polygon
   final List<LatLng> points;
 
   /// The minimum zoom level to fetch tiles for
@@ -43,7 +43,7 @@ class DownloadableRegion {
 
   /// The number of download threads allowed to run simultaneously
   ///
-  /// This will significatly increase speed, at the expense of faster battery drain. Note that some servers may forbid multithreading, in which case this should be set to 1.
+  /// This will significantly increase speed, at the expense of faster battery drain. Note that some servers may forbid multithreading, in which case this should be set to 1, unless another limit is specified.
   ///
   /// Set to 1 to disable multithreading. Defaults to 10.
   final int parallelThreads;
@@ -57,7 +57,7 @@ class DownloadableRegion {
   ///
   /// The checks are conducted by comparing the bytes of the tile at x:0, y:0, and z:19 to the bytes of the currently downloading tile. If they match, the tile is deleted, otherwise the tile is kept.
   ///
-  /// This option is therefore not supported when using satelite tiles (because of the variations from tile to tile), on maps where the tile 0/0/19 is not entirely sea, or on servers where zoom level 19 is not supported. If not supported, set this to `false` to avoid wasting unnecessary time and to avoid errors.
+  /// This option is therefore not supported when using satellite tiles (because of the variations from tile to tile), on maps where the tile 0/0/19 is not entirely sea, or on servers where zoom level 19 is not supported. If not supported, set this to `false` to avoid wasting unnecessary time and to avoid errors.
   ///
   /// This is a storage saving feature, not a time saving or data saving feature: tiles still have to be fully downloaded before they can be checked.
   ///
